requirements.txt
torch
torchvision
timm
opencv-python-headless
matplotlib
Pillow
tqdm
scikit-learn
numpy
gradio